adding license check
