+++ x-pack/plugins/apm/server/feature.ts
@@ -60,3 +60,9 @@ export const APM_FEATURE = {
 
 export const APM_SERVICE_MAPS_FEATURE_NAME = 'APM service maps';
 export const APM_SERVICE_MAPS_LICENSE_TYPE = 'platinum';
+
+export const APM_ML_FEATURE_NAME = 'APM machine learning';
+export const APM_ML_LICENSE_TYPE = 'platinum';
+
+export const APM_CUSTOM_LINKS_FEATURE_NAME = 'APM custom links';
+export const APM_CUSTOM_LINKS_LICENSE_TYPE = 'gold';
