+++ x-pack/plugins/apm/server/routes/settings/custom_link.ts
@@ -3,9 +3,14 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+
+import { i18n } from '@kbn/i18n';
+import Boom from 'boom';
 import * as t from 'io-ts';
 import { pick } from 'lodash';
+import { ILicense } from '../../../../licensing/common/types';
 import { FILTER_OPTIONS } from '../../../common/custom_link/custom_link_filter_options';
+import { APM_CUSTOM_LINKS_FEATURE_NAME } from '../../feature';
 import { setupRequest } from '../../lib/helpers/setup_request';
 import { createOrUpdateCustomLink } from '../../lib/settings/custom_link/create_or_update_custom_link';
 import {
@@ -16,6 +21,18 @@ import { deleteCustomLink } from '../../lib/settings/custom_link/delete_custom_l
 import { getTransaction } from '../../lib/settings/custom_link/get_transaction';
 import { listCustomLinks } from '../../lib/settings/custom_link/list_custom_links';
 import { createRoute } from '../create_route';
+
+function isActiveGoldLicense(license: ILicense) {
+  return license.isActive && license.hasAtLeast('gold');
+}
+
+const INVALID_LICENSE = i18n.translate(
+  'xpack.apm.settings.customizeUI.customLink.forbidden',
+  {
+    defaultMessage:
+      "To create custom links, you must be subscribed to an Elastic Gold license or above. With it, you'll have the ability to create custom links to improve your workflow when analyzing your services.",
+  }
+);
 
 export const customLinkTransactionRoute = createRoute(() => ({
   path: '/api/apm/settings/custom_links/transaction',
@@ -37,6 +54,9 @@ export const listCustomLinksRoute = createRoute(() => ({
     query: filterOptionsRt,
   },
   handler: async ({ context, request }) => {
+    if (!isActiveGoldLicense(context.licensing.license)) {
+      throw Boom.forbidden(INVALID_LICENSE);
+    }
     const setup = await setupRequest(context, request);
     const { query } = context.params;
     // picks only the items listed in FILTER_OPTIONS
@@ -55,9 +75,14 @@ export const createCustomLinkRoute = createRoute(() => ({
     tags: ['access:apm', 'access:apm_write'],
   },
   handler: async ({ context, request }) => {
+    if (!isActiveGoldLicense(context.licensing.license)) {
+      throw Boom.forbidden(INVALID_LICENSE);
+    }
     const setup = await setupRequest(context, request);
     const customLink = context.params.body;
     const res = await createOrUpdateCustomLink({ customLink, setup });
+
+    context.licensing.featureUsage.notifyUsage(APM_CUSTOM_LINKS_FEATURE_NAME);
     return res;
   },
 }));
@@ -75,6 +100,9 @@ export const updateCustomLinkRoute = createRoute(() => ({
     tags: ['access:apm', 'access:apm_write'],
   },
   handler: async ({ context, request }) => {
+    if (!isActiveGoldLicense(context.licensing.license)) {
+      throw Boom.forbidden(INVALID_LICENSE);
+    }
     const setup = await setupRequest(context, request);
     const { id } = context.params.path;
     const customLink = context.params.body;
@@ -99,6 +127,9 @@ export const deleteCustomLinkRoute = createRoute(() => ({
     tags: ['access:apm', 'access:apm_write'],
   },
   handler: async ({ context, request }) => {
+    if (!isActiveGoldLicense(context.licensing.license)) {
+      throw Boom.forbidden(INVALID_LICENSE);
+    }
     const setup = await setupRequest(context, request);
     const { id } = context.params.path;
     const res = await deleteCustomLink({
